dashboard updating properly
